pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "octoclaw"
version = "4.0.0"
description = "Autonomous AI copilot with web admin interface, multi-channel Bot Framework support, and built-in chat. Powered by GitHub Copilot SDK."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "github-copilot-sdk>=0.1.23",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "prompt-toolkit>=3.0",
    "aiohttp>=3.9",
    "aiohttp-session>=2.12",
    "botbuilder-core>=4.14",
    "botbuilder-schema>=4.14",
    "Pillow>=10.0",
    "azure-communication-callautomation>=1.4.0b1",
    "azure-eventgrid>=4.20",
    "azure-identity>=1.19",
    "azure-core>=1.30",
    "azure-keyvault-secrets>=4.8",
    "PyJWT>=2.8",
    "cryptography>=41.0",
    "croniter>=2.0",
]

[project.optional-dependencies]
# voice dependencies are now in main dependencies
voice = []
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "aioresponses>=0.7",
]

[project.scripts]
octoclaw = "octoclaw.cli:main"
octoclaw-bot = "octoclaw.server:main"
octoclaw-admin = "octoclaw.server:main"
octoclaw-env = "octoclaw.env_cli:main"

[tool.setuptools.packages.find]
include = ["octoclaw*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests", "app/runtime/tests"]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]
markers = [
    "slow: marks tests as slow (skipped by default, include with '--run-slow')",
]

[tool.coverage.run]
source = ["octoclaw", "app/runtime"]
omit = [
    "octoclaw/static/*",
    "octoclaw/tests/*",
    "app/runtime/tests/*",
    "app/runtime/static/*",
]
